requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pyjwt==2.8.1
passlib==1.7.4
python-multipart==0.0.6
joblib==1.3.2
numpy==1.26.2
scikit-learn==1.3.2
cors==1.0.1
fastapi-cors==0.0.6
